requirements.txt
-i https://pypi.org/simple
python-dateutil
pytz
requests
scrapy
scrapy-sentry
git+https://github.com/opencivicdata/python-legistar-scraper@854a3725f2325f866f897fbc0a6682f8ab372dc7#egg=legistar
scrapy-wayback-middleware
city-scrapers-core[azure]
pypdf2
freezegun
pytest
flake8
isort
black==19.10b0
